- Changed mc_dwell to double's, because float's have issues with context switches - Changed spinup/spindown of VFD's to something more sensible - Implemented advanced wait for spinup/spindown of VFD's

+++ Grbl_Esp32/src/MotionControl.cpp
@@ -224,7 +224,7 @@ void mc_arc(float*            target,
 }
 
 // Execute dwell in seconds.
-void mc_dwell(float seconds) {
+void mc_dwell(double seconds) {
     if (sys.state == State::CheckMode) {
         return;
     }
